feat: support provide ai
